charts: move formatters to separate file
